Determine winner

+++ contracts/RandomNumberGenerator.sol
@@ -1,10 +1,10 @@
 pragma solidity >=0.4.21 <0.6.0;
 
 contract RandomNumberGenerator {
-    uint8 public randNr = 0;
+    uint32 public randNr = 0;
 
-    function getRandomNumber() public returns (uint8) {
-        randNr = uint8(uint256(keccak256(abi.encodePacked(block.timestamp, block.difficulty)))%251) + 1;
+    function getRandomNumber() public returns (uint32) {
+        randNr = uint32(uint256(keccak256(abi.encodePacked(block.timestamp, block.difficulty)))%251) + 1;
         return randNr;
     }
 }
